ch(structure) - Use app factory.
- Use app factory.
- Introduce blueprints for views.
