+++ run.py
@@ -1,41 +1,7 @@
-import re
-
-from collections import namedtuple
-from datetime import datetime
-
-from flask import Flask, jsonify, request
-from werkzeug.security import generate_password_hash
-
-from db.connection import Connection
-from modals.model import Model
-
-app = Flask(__name__)
-db = Connection.instance()
+from app import create_app
 
 
-@app.route('/signup', methods=['POST'])
-def signup():
-    """Registering a new user"""
-    request_data = request.get_json()
-    email = request_data['email']
-    reg = re.match(
-        '([a-zA-Z0-9]+)([.]*)([-]*)([a-zA-Z0-9]+)@{1}([a-z]+)\.([a-z]+)',
-        email)
-    if reg is None:
-        return jsonify(message="Invalid email address"), 401
-    password = request_data['password']
-    # specify columns to insert data into
-    Columns = namedtuple('Columns', 'email password date_created')
-    new_user = Columns(
-        email=email,
-        password=generate_password_hash(password),
-        date_created=str(datetime.now())
-        )
-    user_table = Model(db, 'user_table')
-    user_table.create(new_user)
-    return jsonify(
-        message="Registration Successful"
-        ), 201
+app = create_app()
 
 if __name__ == '__main__':
     app.run(debug=True)
